[field] Move components to "components" directory
