+++ packages/@sanity/field/src/diff/annotations/index.ts
@@ -1,7 +1,2 @@
 export * from './hooks'
 export * from './helpers'
-export * from './DiffAnnotation'
-export * from './DiffAnnotationCard'
-export * from './DiffAnnotationTooltip'
-export * from './DiffAnnotationTooltipContent'
-export * from './AnnotatedStringDiff'
