--- packages/@sanity/field/src/diff/changes/index.ts
@@ -1,3 +1,0 @@
-export * from './ChangeResolver'
-export * from './DocumentChangeContext'
-export * from './ChangeList'
