+++ packages/@sanity/field/src/diff/components/AnnotatedStringDiff.css
