+++ packages/@sanity/field/src/diff/components/DiffAnnotationTooltip.tsx
@@ -1,8 +1,8 @@
 import {Tooltip} from 'part:@sanity/components/tooltip'
 import React, {createElement} from 'react'
+import {getAnnotationAtPath} from '../annotations'
 import {DiffAnnotationTooltipContent} from './DiffAnnotationTooltipContent'
 import {AnnotationProps, AnnotatedDiffProps} from './DiffAnnotation'
-import {getAnnotationAtPath} from './helpers'
 
 interface BaseAnnotationProps {
   as?: React.ElementType | keyof JSX.IntrinsicElements
