+++ packages/@sanity/field/src/diff/components/index.ts
@@ -1,3 +1,11 @@
-export * from './ChangeLayout'
-export * from './MetaInfo'
+export * from './AnnotatedStringDiff'
 export * from './Change'
+export * from './ChangeLayout'
+export * from './ChangeList'
+export * from './ChangeResolver'
+export * from './DiffAnnotation'
+export * from './DiffAnnotationCard'
+export * from './DiffAnnotationTooltip'
+export * from './DiffAnnotationTooltipContent'
+export * from './DocumentChangeContext'
+export * from './MetaInfo'
