+++ packages/@sanity/field/src/diff/index.ts
@@ -1,6 +1,6 @@
 export * from '../types'
+
 export * from './annotations'
-export * from './changes'
-export * from './resolve'
 export * from './components'
 export * from './helpers'
+export * from './resolve'
